# Fix overcompilation with scalac -release by dummy mapping ct.sym entries