--- a/internal/zinc-core/src/main/scala/sbt/internal/inc/MappedVirtualFile.scala
+++ b/internal/zinc-core/src/main/scala/sbt/internal/inc/MappedVirtualFile.scala
@@ -58,8 +58,12 @@ class MappedFileConverter(rootPaths: Map[String, Path], allowMachinePath: Boolea
       case Some((key, rootPath)) =>
         MappedVirtualFile(s"$${$key}/${rootPath.relativize(path)}".replace('\\', '/'), rootPaths)
       case _ =>
+        def isCtSym =
+          path.getFileSystem
+            .provider()
+            .getScheme == "jar" && path.getFileSystem.toString.endsWith("ct.sym")
         def isJrt = path.getFileSystem.provider().getScheme == "jrt"
-        if (isJrt || path.getFileName.toString == "rt.jar")
+        if (isJrt || path.getFileName.toString == "rt.jar" || isCtSym)
           DummyVirtualFile("rt.jar", path)
         else if (allowMachinePath) MappedVirtualFile(s"$path".replace('\\', '/'), rootPaths)
         else sys.error(s"$path cannot be mapped using the root paths $rootPaths")